fixed small merge conflict

+++ src/Engine/Systems/Rendering/MeshList/MeshList.hpp
@@ -43,7 +43,7 @@ public:
 
     // inline wrapper functions for basic list functionality
 
-    inline size_t size() { return (int)list.size(); }; // get the number of gameObjects in the meshList
+    inline size_t size() { return list.size(); }; // get the number of gameObjects in the meshList
 
     inline GameObject * quickGet( int i ) { return list[i]; } // replacement for accessing elements in the list. only accesses the element, and does nothing else
